Pre develoment

+++ backend/src/controllers/usersController.js
@@ -2,17 +2,33 @@ import * as userService from '../services/userService.js';
 
 export const createUser = async (req, res) => {
   try {
-    const newUser = await userService.createUser(req.body);
+    const { nombre, email, contrasena } = req.body;
+    if (!nombre || !email || !contrasena) {
+      return res.status(400).json({ success: false, message: 'Faltan datos requeridos' });
+    }
+
+    const newUser = await userService.createUserService({ nombre, email, contrasena });
     res.status(201).json({ success: true, data: newUser });
   } catch (error) {
-    res.status(400).json({ success: false, message: error.message });
+    console.log(error);
+    next(error);
   }
 };
 
 export const updateUser = async (req, res) => {
   try {
     const { userId } = req.params;
-    const updatedUser = await userService.updateUser(userId, req.body);
+    const { id_usuario } = req.user;
+
+    // Verificar si el usuario tiene permiso para actualizar
+    if (String(userId) !== String(id_usuario)) {
+      return res.status(403).json({
+        success: false,
+        message: 'No tienes permiso para actualizar este usuario',
+      });
+    }
+
+    const updatedUser = await userService.updateUserService(userId, req.body);
     if (!updatedUser) {
       return res
         .status(404)
@@ -20,13 +36,22 @@ export const updateUser = async (req, res) => {
     }
     res.status(200).json({ success: true, data: updatedUser });
   } catch (error) {
-    res.status(400).json({ success: false, message: error.message });
+    next(error);
   }
 };
 
 export const deleteUser = async (req, res) => {
   try {
     const { userId } = req.params;
+    const { id_usuario } = req.user;
+
+    // Verificar si el usuario tiene permiso para actualizar
+    if (String(userId) !== String(id_usuario)) {
+      return res.status(403).json({
+        success: false,
+        message: 'No tienes permiso para actualizar este usuario',
+      });
+    }
     const result = await userService.deleteUser(userId);
     if (!result) {
       return res
@@ -37,6 +62,6 @@ export const deleteUser = async (req, res) => {
       .status(200)
       .json({ success: true, message: 'Usuario eliminado exitosamente' });
   } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
+    next(error);
   }
 };
